feat(habitat): move hook orchestration into service module
Move habitat hook command orchestration into the owned Effect-oRPC hook service module and delete the obsolete src/lib/hooks.ts aggregate wrapper.

Refuse the old runHook helper and lib/hooks aggregate D0 surfaces instead of preserving compatibility wrappers. Lower-level hook-runtime provider/resource drainage remains a follow-on packet.
