+++ openspec/changes/deep-habitat-effect-substrate-architecture/workstream/public-contract-risk-register.md
@@ -17,7 +17,7 @@ is invalid without that ledger.
 | `CheckReport` v1 | `src/lib/check-report.ts`; `src/lib/diagnostics.ts`; `src/lib/check/schema.ts`; root exports | Internal domain rename can leak into JSON fields, especially `ownerTool`, duration, baseline fields | Golden schema validation and compatibility mapping from new internals to v1 |
 | Classify output | `src/lib/classify-core/**`; `src/lib/classify.ts`; D0 matrix classify rows | Nx provider split can change project-path/diff/refusal states | H8 classify path/diff matrix and JSON/text rendering parity |
 | Verify receipts | `src/lib/verify/**`; `src/commands/verify.ts` | Git/Nx provider split can change receipt shape, timestamps, cache task fields, or base resolution language | Receipt schema tests, fake Git/Nx provider matrix, command JSON parity |
-| Hook behavior | `.husky/pre-commit`; `.husky/pre-push`; `src/lib/hooks.ts`; `src/lib/hook-runtime/**`; `src/commands/hook.ts` | Effect resource scoping can change staged/restage behavior or local-only non-claims | staged/unstaged/partial/restage tests, pre-push base proof, hook non-claim receipt |
+| Hook behavior | `.husky/pre-commit`; `.husky/pre-push`; `src/service/modules/hook/**`; `src/lib/hook-runtime/**`; `src/commands/hook.ts` | Effect resource scoping can change staged/restage behavior or local-only non-claims | staged/unstaged/partial/restage tests, pre-push base proof, hook non-claim receipt |
 | Package exports | `tools/habitat-harness/package.json` exports `.` and `./plugin`; `src/index.ts`; `src/plugin.ts` | Moving internals can break consumers currently importing broad root exports | Public-surface facade packet with callsite census and D0 compatibility matrix update |
 | Nx plugin and generators | `src/plugin.ts`; `src/plugin/**`; `src/generators/**`; `generators.json`; package `generators` field | Domain extraction can break Nx generator contract, schema output, or plugin target metadata | generator tests, schema generation check, plugin metadata/typecheck |
 | Root scripts and Nx targets | `package.json` scripts; root AGENTS tooling defaults; package Nx targets | Provider command materialization can bypass repo-local pinned tools or dependency ordering | root script parity, local package script checks, Nx target proof where packet touches targets |
