+++ openspec/changes/deep-habitat-effect-substrate-architecture/workstream/source-movement-map.md
@@ -49,7 +49,7 @@ find tools/habitat-harness/src -type f | sort
 | `src/lib/git-state.ts` | 1 | `src/providers/git/**` | move | Git status/head/diff facts become provider operations. |
 | `src/lib/graph.ts` | 1 | `src/domains/workspace-graph-integration/**`, `src/resources/temp-dir.ts` | split | Temp lifecycle and graph read/write move out of domain logic. |
 | `src/lib/verify/**` | 7 | `src/domains/proof-contract/**`, `src/domains/workspace-graph-integration/**`, `src/providers/nx/**`, `src/providers/git/**` | split | Receipts remain public-contract guarded; affected execution becomes provider/domain composition. |
-| `src/lib/hooks.ts` | 1 | `src/domains/local-feedback/hooks.ts` | move | Public hook behavior remains command-facing; mutable staged/restage logic moves into local feedback domain. |
+| `src/lib/hooks.ts` | 1 | `src/service/modules/hook/**`, later `src/domains/local-feedback/**` | split and delete | `runHook` package helper export is refused; hook command orchestration moves to the Effect-oRPC service module, and lower-level staged/restage/provider ownership drains in follow-on packets. |
 | `src/lib/hook-runtime/**` | 11 | `src/domains/local-feedback/**`, `src/providers/{git,biome,grit,nx,husky}/**`, `src/resources/write-set.ts` | split | Optional runtime dependency bag is replaced by Effect requirements and fake layers. |
 | `src/lib/pattern-apply/**` | 11 | `src/domains/transformation-transaction/**`, `src/providers/grit/**`, `src/domains/pattern-governance/**` | split | Apply admission, worktree, rollback, record, and render become transaction/domain responsibilities. |
 | `src/lib/protected-zones/**` | 9 | `src/domains/protected-zone-authority/**`, `src/resources/filesystem.ts` | move | Protected write decisions are domain-owned; fs mutation is resource/provider-owned. |
