+++ tools/habitat-harness/src/service/modules/hook/context.ts
@@ -1,0 +1,9 @@
+import type { HookRuntime } from "../../../lib/hook-runtime/runtime.js";
+
+export interface HookServiceOptions {
+  runtime?: HookRuntime;
+}
+
+export interface HookServiceContext {
+  readonly hook?: HookServiceOptions;
+}
