+++ tools/habitat-harness/src/service/modules/hook/router.ts
@@ -2,7 +2,7 @@ import { module as hookModule } from "./module.js";
 import { runHookService } from "./run.js";
 
 export const hookRouter = {
-  run: hookModule.run.effect(({ input }) => runHookService(input)),
+  run: hookModule.run.effect(({ context, input }) => runHookService(input, context.hook)),
 };
 
 export const router = hookRouter;
